change float to double

+++ analyzer/src/main/java/ru/neoflex/scammertracking/analyzer/domain/model/Coordinates.java
@@ -3,7 +3,7 @@ package ru.neoflex.scammertracking.analyzer.domain.model;
 import java.io.Serializable;
 
 public class Coordinates implements Serializable {
-    public Coordinates(float latitude, float longitude) {
+    public Coordinates(double latitude, double longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
@@ -11,22 +11,22 @@ public class Coordinates implements Serializable {
     public Coordinates() {
     }
 
-    private float latitude;
-    private float longitude;
+    private double latitude;
+    private double longitude;
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 }
